Rename validators.SettingValueValidator to constraints.SettingConstraint
